Feat(RacingCar) : 자동차 경추 게임 우승자 개발(4단계 피드백)
- 사용하지 않는 테스트 코드 삭제
- 클래스에 대한 클래스Test로 분리
